Fixed bug in the sensor editor
Fixed the issue where the editor wouldn't show the 'target int' field for local integer animator parameters whenever multiple animators with different parameters were assigned to an avatar. This issue occurred because the list of local parameters may not be identical to the options displayed in the editor, causing the editor to store a different local parameter.